--- a/activities/models/timeline_event.py
+++ b/activities/models/timeline_event.py
@@ -62,6 +62,7 @@ class TimelineEvent(models.Model):
             # This relies on a DB that can use left subsets of indexes
             ("identity", "type", "subject_post", "subject_identity"),
             ("identity", "type", "subject_identity"),
+            ("identity", "published"),
         ]
 
     ### Alternate constructors ###