Rendering the home timeline is a fairly intense query, and is used in several of the heaviest APIs in the system. Currently they are sorted by -published which is not indexed, so the query planner tries various ways to get the data as best as possible.

With an explicit index though, it can readily grab the most recent events, and move on.

Example stats from my takahe.tabletcorry.com server on my user, rendering the home timeline:

Before:
     Planning Time: 6.120 ms
     Execution Time: 12.109 ms
After:
    Planning Time: 4.797 ms
    Execution Time: 0.769 ms

This improvement is not super visible in benchmarking due to other noise dominating, but it will doubtless reduce load on the DB overall.